Blog List Expansion 4.15-4.23

+++ controllers/blogsController.js
@@ -1,33 +1,70 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const config = require('../utils/config')
+const middleware = require('../utils/middleware')
+
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
       response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const blog = await new Blog(request.body)
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const blog = request.body
 
-  if (!blog.title || !blog.url) {
+  const decodedToken = jwt.verify(request.token, config.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  
+  const user = request.user
+
+  const newBlog = new Blog(
+    {
+      title: blog.title,
+      author: blog.author,
+      url: blog.url,
+      likes: blog.likes,
+      user: user.id
+    }
+  )
+
+  if (!newBlog.title || !newBlog.url) {
     response.status(400).end()
   } else {
-    const result = await blog.save()
-    response.status(201).json(result)
+    const savedBlog = await newBlog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.status(201).json(savedBlog)
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const decodedToken = jwt.verify(request.token, config.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(403).json({ error: 'you do not have permission to delete this blog' })
+  }
 })
 
-blogsRouter.put('/:id', async (request, response) => {
-  const blog =   {
-    'title': request.body.title,
-    'author': request.body.author,
-    'url': request.body.url,
-    'likes': request.body.likes
+blogsRouter.put('/:id', middleware.userExtractor, async (request, response) => {
+  const blog =  {
+    title: request.body.title,
+    author: request.body.author,
+    url: request.body.url,
+    likes: request.body.likes
   }
 
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
